Slice the coverage C tests by count and drop the duration machinery
* The three C legs each computed their own duration-balanced slice, which is a
  partition only if all three weigh the tests identically. They restored the
  duration bundle through a restore-keys prefix, which resolves to whichever
  entry is newest when that leg starts, and the legs do not start together, so
  a bundle saved by a concurrently finishing run gave later legs different
  weights and three incompatible partitions
* Seen on run 33457691603: 84 selections covering 67 unique tests, so 17 ran
  twice and 16 not at all. codecov read that as -3.72% with whole test files at
  zero hits, which looked like a regression in the PR under review
* Use meson's own --slice instead. Measured over the current suite that costs
  115s -- a 518s worst slice against 403s -- on legs that take 9-11 min while
  this job's critical path is py-xcor at 30-47 min. When one test grows enough
  for that to matter the answer is to split it, which is worth doing anyway
* Removes the duration cache, its version knob, the per-leg extract and upload
  steps, and the merge-test-durations job. test_slicer.py loses plan and
  extract, leaving only the timing report, so it becomes test_summary.py
* Net 98 lines removed from the workflow, and the failure mode is gone rather
  than guarded: there is no longer a per-leg input that can disagree
